feat: glob import plugin (#1510)
<!-- Thank you for contributing! -->

### Description
The pr intend to implement Vite GlobImportPlugin, doc is
https://vitejs.dev/guide/features.html#glob-import.

Before we discussed using `transform` hook return `ast` to do it, it
can't implement because we always need to re-parse and re-codegen for
it. So here add `transform_ast` hook to do this, if you any ideas, I'm
feel free using `transform` hook to do it.
<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->