chore(deps): bump oxc to `0.19.0` (#1560)
You can read about the new builder methods here:
https://github.com/oxc-project/oxc/pull/3890

We also provide the `IntoIn` and `FromIn` in the `oxc_allocator` crate
so I've removed the rolldown helper's version.

I didn't touch the snippets but now some of them can painlessly be
replaced with builder methods since we accept generic arguments for
things that can be allocated on the arena. So now you can pass `&str`
where `Atom` or `&'ast str` is expected. Or pass in a stack ast type and
have it boxed in the builder using `IntoIn` bounds.