refactor: rename `RenderedChunk` to `RollupRenderedChunk` (#1606)
<!-- Thank you for contributing! -->

### Description

The design of rolldown is to be a self-consistent bundler. All the API
it provides it are due to, as a bundler, it have to provide. And the
compatibility with rollup are also supported by these API, we don't need
come up with something that is specific for rollup.

However, in practice, this's not possible, we have to add something that
just for rollup compatibility and I want to know how much code we are
added due to this purpose.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->