chore: add minify benchmark (#1667)
<!-- Thank you for contributing! -->

### Description

The pr add the minify benchmark and minify-sourcemap benchmark. Here
remove `scan-xx-sourcemap` benchmark, because it's sourcmap is
unmeaning, it will not generate sourcemap.

Also add it to see for https://github.com/rolldown/rolldown/pull/1664. i
think render chunk remapping case is small..

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->