feat: support `output.exports` option. (#1680)
### Description

resolves #1628

This pull request addresses the `output.exports` option related to
[Rollup](https://rollupjs.org/configuration-options/#output-exports).

The PR does not yet address deconfliction; I will implement it if
needed. (Rollup uses `getExportMode`, which is `determine_export_mode`
in this PR, during deconfliction).

### Details

1. Added options in both the Rust and TypeScript modules.
2. Introduced the function `determine_export_mode` to handle the `auto`
value. It is adapted from Rollup but uses `match` instead of an `if`
chain.
3. Managed `export_mode` in
`crates/rolldown/src/utils/chunk/render_chunk_exports.rs`.

### Issues Waiting for Fixes in this PR

1. Remove the duplicated enums `ExportMode` and `OutputExports`, even
though `OutputExports` includes `Auto` while `ExportMode` does not.
2. Improve support for the `iife` format. The `iife` format is still
unstable in this project, and it's uncertain if it can handle the
`output.exports` issue for `iife`. However, a prototype has been
provided.
3. Used `anyhow::bail` for error handling. The error messages are not
clear enough and need improvement. Contributors are welcome to address
this, and `TODO` comments have been added to the relevant sections.

This is the first time I have contributed to the core code. I am willing
to fix problems in this PR promptly! :-)

---------

Co-authored-by: underfin <likui.underfin@gmail.com>