feat: support `output.exports` for `IIFE` format align with rollup. (#1691)
In PR #1680, support for `output.exports` was only available for the
`Cjs` format. This PR extends the support to the `IIFE` format.

### Modifications

1. **Wrapper Function Update:**
- In `crates/rolldown/src/ecmascript/format/iife.rs`, the declaration
part of the wrapper function has been rewritten. The `exports` argument
is now enabled only if `output.exports` is `named`, and the non-name
mode issue has been fixed. Similar changes have been made to the ending
part of the wrapper function.

2. **Return Value Alignment:**
- In `crates/rolldown/src/utils/chunk/render_chunk_exports.rs`, the
value is now returned directly instead of assigning it to `exports`,
aligning with Rollup’s behavior when `output.exports` is `default` and
the `format` is `iife`.

3. **Removal of `__esModule` Property:**
- The statement `Object.defineProperty(exports, '__esModule', { value:
true });` has been removed when `output.exports` is `default` and the
`format` is either `cjs` or `iife`, as it previously caused errors in
some cases.

4. **Testing:**
- Added tests for the `iife` format to ensure correct handling of
`output.exports`.

These changes ensure better alignment with Rollup and improve the
robustness of the `IIFE` format support in the project.