feat: support loading dataurl as module. (#1741)
This PR ports functionality from
[Vite](https://github.com/vitejs/vite/blob/main/packages/vite/src/node/plugins/dataUri.ts)
and enables esbuild's tests.

Please note that there is an outstanding issue referenced
[here](https://github.com/evanw/esbuild/blob/main/internal/bundler_tests/snapshots/snapshots_loader.txt#L602)
that is pending a fix. However, it seems unlikely that this PR will
address it, as the complexity of handling the module as an external
module again in `load_source.rs` is considerable.

---------

Co-authored-by: Yunfei <i.heyunfei@gmail.com>