feat: support `output.esModule` option (#1812)
This PR introduces support for namespace blocks in IIFE or CJS format
outputs, treating them as ES Modules instead of relying solely on
`output.exports` being `named`. @underfin (I'm not in a hurry to work on
this as I will likely spend a lot of time on my competition project next
month. :)

### Modifications

1. Removed redundant usage of `determine_export_mode`.
2. Added `render_namespace_block`, inspired by Rollup, but implemented
based on behavior observed in the REPL.
3. Added corresponding tests to validate the changes.

### Behaviors

1. **Namespace block is disabled if `output.exports` isn't `named`.**
2. **Namespace block is disabled if `output.format` is `esm` or `app`.**
   - `esm` should always be an ES module.
   - `app` would never be a module.
3. **Entry Point with Default Export:**
- When `output.esModule` is set to `'always'` (or `true` for Rollup
compatibility), or `'if-default-prop'` (similar to Rollup), it will
render `Object.defineProperty(exports, '__esModule', { value: true })`.
   - Left an entry for `generatedCode.symbol` for future updates.

### Rationale for Not Using `boolean | 'if-default-prop'`

During the conversion from JavaScript to Rust, the program encountered a
panic when the source value was a boolean. Overriding `true` or `false`
with a string was considered inelegant, so this approach was not used.

---------

Co-authored-by: underfin <likui.underfin@gmail.com>