feat(node): support passing `custom` in `PluginContext#resolve` (#1874)
<!-- Thank you for contributing! -->

### Description

The pr intend to support `PluginContext#resolve#option#cutsom`, it will
be transfer to `ResolveId#ResolveOption#custom` to do some plugin
resolve logic.

It is used at vite resolver, and the struct is complex, see
[here](https://github.com/vitejs/vite/blob/main/packages/vite/src/node/plugins/resolve.ts#L181).
We can't directly convert it to rust struct. So here store a index to
reference the js side custom options. It is difficult to expand, but i
think it's less usage at rust side plugins, we could improve it if we
have demand.

Other things, here fixed a logic about `PluginContextData`, it should be
singleton, But it make complex to shared data at parallel worker plugin,
We could leave it at now util we find a good idea.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->