feat: provide more friendly diagnostics if could not read path from a resolveId hook (#1866)
<!-- Thank you for contributing! -->

### Description

Closes #1851
<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->

I don't know if it should be simply regarded as a `rolldown internal
error`.

At first, I tried to implement the incorrect format of `esbuild`, but
later I found that `rolldown` is not enough to support me in doing so.

![wecom-temp-17469-893230717fe496ece769b4a6e33e9aeb](https://github.com/user-attachments/assets/7cd2efdc-bfe5-48ef-aec6-debd82b1e402)

Therefore, I attempted to implement the incorrect format of `rollup`.

![wecom-temp-92390-a41e8d41a7c53c93daa4c7e305564416](https://github.com/user-attachments/assets/61f184d1-7e46-4182-957a-a48982238cf3)

Due to the current lack of support for plugins in `rolldown testing`, I
am unable to conduct testing on `rust`. In addition, testing in
`fixtures` can lead to `build errors`, so currently I can only ignore
it.
<img width="700" alt="image"
src="https://github.com/user-attachments/assets/41e3378c-4a25-45ae-b991-8a08cd785eff">