refactor(rust/plugin): set `ModuleTable` to `PluginContext` in place (#1967)
<!-- Thank you for contributing! -->

### Description

This PR doesn't reduce the risk of using `unsafe` compared to the old
code.

The old code requires every fields of `PluginDriver` to be `Clone`, it
means we have to wrap ever fields(for un-`Clone` type) with `Arc` or do
a full `Clone.

This `Arc<Many Arc<...>>` patterns doesn't feel right and this PR solve
it.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->