feat: support namespaced name and `output.extend` for IIFE format. (#1861)
This pull request supports the use of namespaced names (including the
dot character) and the `output.extend` option for IIFE format aligning
with Rollup. Tests are included, and further refactoring is planned (not
included in this PR, as it is better to conduct refactoring when more
formats are supported, in my opinion).

---------

Co-authored-by: underfin <likui.underfin@gmail.com>