feat(node): support CLI options. (#1963)
resolves #1955.

This public relations campaign leverages the `parseArgs` function
provided by the `node:util` module and the JSON schema generated by the
`zod` library to automatically generate CLI options. This approach
enhances the elegance and convenience of the process.

In my opinion, the runtime-generated schema is performance-degrading.
While it is feasible to generate it during compilation via a plugin, it
is currently too complex for this particular pull request, and also
requires support `TypeScript` for config file.