fix: ensure ordering of assign chunk name (#1981)
<!-- Thank you for contributing! -->

### Description

The esbuild using `Chunk#bits` to sorted chunks, but the order of
`Chunk#bits` is not stable, eg `BitSet(0) 00000001_00000000` >
`BitSet(8) 00000000_00000001`. It couldn't ensure the order of dynamic
chunks and common chunks.
Consider the compare `Chunk#exec_order` should be faster than
`Chunk#bits`, we use `Chunk#exec_order` to sort chunks.
 Note Here could be make sure the order of chunks.
 - entry chunks are always before other chunks
 - static chunks are always before dynamic chunks
 - other chunks has stable order at per entry chunk level

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->