feat(node/composingJsPlugins): support `resolveId` hook (#2021)
<!-- Thank you for contributing! -->

### Description

You could actually call `this.resolveId` in `buildStart`... hooks, so
the current bailout logic is wrong. To get a correct semantic, we
probably need to bailout all hooks that could call `this.resolveId`.
This basically make `composingJsPlugins` unuseable.

However, this PR solve the issue partially, which means, with this PR,
we could ensure the correct semantic of calling `this.resolve` in
`resolveId` hook.

Fixes for other hooks would be raised later. Basically we need intercept
all hooks's `PluginContext` for plugins that implemented `resolveId`
hook.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->