perf: use `simdutf8` to validate UTF-8 when reading files (#2040)
### Description

I'm not familiar with Rolldown's codebase, so I'm not sure whether this
function is the main one used to read source text from files - so this
may be in wrong place/not enough places.

`std::fs::read_to_string` internally calls `std::str::from_utf8` to
check that the file is valid UTF-8. `str::from_utf8` is not so efficient
as it doesn't use SIMD. `simduft8` should be faster.

I assume that this difference won't show up on benchmarks as they likely
won't include IO. But I'd hope it'd have an effect in the real world.
How would we go about testing that assumption? Are there any benchmarks
which do exercise this function?