feat: support `output.inlineDynamicImports` option. (#2081)
resolves #1841.

The logic for the `inline dynamic imports` option has been implemented,
but it is currently hard-coded only for `iife` format. I have added an
option called `inlineDynamicImports` that aligns with Rollup’s approach
to manually handle this scenario.

Rollup’s default value for this option is
[`false`](https://rollupjs.org/configuration-options/#output-inlinedynamicimports).
However, I believe we can check if the format is `iife` to determine
whether to enable this option, but it may be not so elegant. Any
suggestions on how to proceed?

---------

Co-authored-by: underfin <likui.underfin@gmail.com>