chore(rust): rename `PreliminaryAsset` to `InstantiatedChunk` (#2097)
<!-- Thank you for contributing! -->

### Description

We need a concept to represent the thing between `Chunk` and `Asset`.

It's used to be `PreliminaryAsset`, but it's not proper with considering
terms in rollup. For example, we now pass `PreliminaryAsset`s to
`render_chunk` hook. This is quite confusing.

Rendering chunks to assets is a process that turn abstraction to
something specific, which makes `InstantiatedChunk` a perfect term to
represent the middle product between chunks and assets.

So the flow is now:

`Chunk` ---different generators---> `InstantiatedChunk`s -> `Asset`.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->