fix:  dicable dce when disable tree shake (#2335)
<!-- Thank you for contributing! -->

### Description
It would be better to disable DCE for this test, so can see the actual
output from replace plugin, but I couldn't figure our how to do that. In
Rollup, `treeshake: false` does that, but `treeshake:
TreeshakeOptions::Boolean(false)` seemed to have no effect.

_Originally posted by @overlookmotel in
https://github.com/rolldown/rolldown/pull/2322#discussion_r1775296379_
            
<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->