refactor: watcher event channel msg (#2493)
<!-- Thank you for contributing! -->

### Description

Before the windows ci failed by
`tx.lock`(https://github.com/rolldown/rolldown/actions/runs/11456512537/job/31874824001),
i changed it to `tx.blocking_lock`, but it make the `Watcher#invalidate`
logic not work, the behavior will be changed once build once. I try to
add test for it, but it is difficult because the different platforms
emit change event is not stable.


Here refactor it to avoid adding `Mutex` to `tx` by using
`sync::mpsc::channel`, but it need to customer define close event.

Here other things is stop thread when channel close.


<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->