refactor(error): remove `UnhandleableResult` (#2501)
<!-- Thank you for contributing! -->

### Description

We used to try to strongly distinguish two kinds of error.

- Program error, not related to bundling, such as convert u64 to usize
in u32-based platform
- Business error, such as importing a file which doesn't exist while
bundling.

After a long time practicing, these two kinds of error aren't show much
differences for writhing web bundler. In fact, they share a same and
important characteristic which is to break the control flow.

So we don't need to make strong distinguish anymore and they are all
treated and handled by the same error mechanism in rolldown.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->