fix: make sure the sort compare function satisfy `total order` (#2621)
<!-- Thank you for contributing! -->

### Description
according to the doc: 
> If the comparison function `compare` does not implement a [total
order](https://en.wikipedia.org/wiki/Total_order) the resulting order
of elements in the slice is unspecified. All original elements will
remain in the slice and
any possible modifications via interior mutability are observed in the
input. Same is true
if `compare` panics.
1. Closed https://github.com/rolldown/rolldown/issues/2616
## Ref

https://blog.rust-lang.org/2024/09/05/Rust-1.81.0.html#new-sort-implementations
<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->