feat: support PluginContext.load (#2687)
<!-- Thank you for contributing! -->

### Description

Support `PluginContext.load`, ref
https://rollupjs.org/plugin-development/#this-load. close
https://github.com/rolldown/rolldown/issues/2355.

The pr using msg channel to hand extra modules need to load. Here need
to use one channel to avoid
https://github.com/rolldown/rolldown/pull/2687#discussion_r1836055076.

> Here has a another way to do it. Using `load_via_plugins`, like
`resolve_id_via_plugins`, but it need to a lot of change the relation
crate dependencies and meet a cycle dependencies problem at moving
`create_ecma_view`, it is difficult to resolve. try here at
https://github.com/rolldown/rolldown/pull/2690.

The pr implements it as much as possible, but it also has a problem,
consider some cases at here.
- `ModuleTables` has not the module `A`, here only fetch it.
- `ModuleTables` has module `A` and already fetched, here return already
fetched result, using `PluginContext#modules`.
- `ModuleTables` has module `A` but it is not fetched finished, here may
lost result because here can't ensure the fetch time before the module
task is finished. The case maybe happend because it is run at parallel.
We could re-fetch it at the case, but it call the module-related hook
again, It could be put as low priority to resolve.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->