fix: align resolve behavior of `new URL(..., import.met.url)` with webpack (#2738)
<!-- Thank you for contributing! -->

### Description

So the behavior is for bare identifier, bundler will resolve it as
relative path first instead of npm modules. It means that `foo` would be
resolve to `./foo.js` first. It'll hide the npm module `foo`.

More contexts:

- To prevent the above behavior, parcel supports [`npm:xxx`
pattern](https://parceljs.org/blog/v2/#new-npm%3A-scheme) to force
resolve to node modules first. Eg: `new URL("npm:foo",
import.meta.url)`.

This behavior is not supported by rolldown. We'll wait users feedbacks
and see if we need to support this feature.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->