docs(contrib): add coding style for naming in rust (#2765)
<!-- Thank you for contributing! -->

### Description

The naming style is not a strict rule, as the book said

> These guidelines should not in any way be considered a mandate that
crate authors must follow, though they may find that crates that conform
well to these guidelines integrate better with the existing crate
ecosystem than those that do not.

It's ok to violate these suggestions, but you should have a reasonable
explanation, such as the book didn't cover this pattern, instead of
personal habit.

- I'm open to other styles. But following
https://rust-lang.github.io/api-guidelines/ should be more pratical
choice than others due to

> They are authored largely by the Rust library team, based on
experiences building the Rust standard library and other crates in the
Rust ecosystem.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->