feat: don't normalize options on js side - part2 (#2848)
### Description

I found that removing `normalizeInputOptions` is a bit tricky because it
is tightly coupled with the `Plugin` on the js side.

I reviewed the context of issue #1041, but I'm still unsure if I'm on
the right track. Therefore, I'm submitting this PR to ensure that what
I'm doing is correct. I would appreciate it if you could point out the
next steps I should take.


> I also think about it before, we could using original `external`
option to improve it. If the user not changed normalized `external`
option, we just using original `external` option to binding option, it
could be `array|function`.
> 
> 
> 
> I think about this too. But it requires us to change the definition of
`NormalizedInputOptions`. If this is acceptable, we could just use
`InputOptions` as the `NormalizedInputOptions`. I kind of feel that this
makes the normalization step in js side meaningless, so I didn't propose
this way.
_Originally posted by @hyf0 in
[#1041](https://github.com/rolldown/rolldown/issues/1041#issuecomment-2098313025)_