fix: ensure transform hook return vaild sourcemap to binding (#2939)
<!-- Thank you for contributing! -->

### Description

The `transform` hook maybe return an invalid sourcemap, the added test
is an example, it used the `sourceIndex`, but not the `sources` field.
The pr fixed it to avoid failed at
`rolldown_sourcemap::SourceMap::from_json`.

The rust side already has the feature at
`crates/rolldown_plugin/src/plugin_driver/build_hooks.rs#transform`, it
also is valid if the rust plugins provider an invalid sourcemap. Maybe
we could remove it at rust side if we could ensure a valid sourcemap, i
think it could be preserved at now because it is sugar to usage.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->