feat: expose app format option for hmr prototype (#2994)
### Description

Currently I'm trying a different approach from hmr-poc branch for
implementing hmr. For this approach, for now I only need isolating
module finalizer logic, so exposing current app format is sufficient to
make prototype as in https://github.com/hi-ogawa/vite/pull/6.

Since this PR's change is fairly minimal, I'm hoping that we can merge
this to main, so that I can experiment further on the nightly release
and iterate on it.

---

For the context, the idea I'm exploring is to give up `HmrModuleLoader`
for now as I see some limitations (see
https://github.com/rolldown/rolldown/issues/2866), but instead, always
do a full build and extract hmr chunk by comparing `chunk.modules`
between two builds, which is fortunately available for js api.