feat(watcher): support multiply watch options (#3037)
<!-- Thank you for contributing! -->

### Description

The `Bundler::watch` api is not suitable for watcher if it is multiply
watch options, so here add a new `BindingWatcher`. The binding the
multiply `Bundler` to create `BindingWatcher` is impossible, so here
pass binding multiply options to create it.

The rollup is create `chokidar` for each `Bundler`, it has more overhead
at rolldown, because we need to create thread to listen event. So the
`notify` option is used for all `Bundler`, i will check it at next.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->