feat: give warning for multiply notify option (#3085)
<!-- Thank you for contributing! -->

### Description

related https://github.com/rolldown/rolldown/pull/3037.

The rollup is create chokidar for each Bundler, it has more overhead at
rolldown, because we need to create thread to listen event. So the
notify option is used for all Bundler, here give warning for multiply
notify options.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->