refactor: use safer logic to get `ModuleTable#modules` (#3156)
### Description

The `importers` and `modules` in `intermediate_normal_modules` are
mapped one-to-one by `idx`. If a module fails to be fetched during the
`fetch_all_modules` process, `flatten()` will break the corresponding
mapping.


https://github.com/rolldown/rolldown/blob/ddbe0a9e8252f61380a4a5b57d82e32a00d0791b/crates/rolldown/src/module_loader/module_loader.rs#L407-L414

https://github.com/rolldown/rolldown/blob/ddbe0a9e8252f61380a4a5b57d82e32a00d0791b/crates/rolldown/src/module_loader/module_loader.rs#L418

Although most cases are already handled by the code below:

https://github.com/rolldown/rolldown/blob/ddbe0a9e8252f61380a4a5b57d82e32a00d0791b/crates/rolldown/src/module_loader/module_loader.rs#L387-L389

What I’ve done is simply make the code safer. If my understanding is
incorrect, feel free to close this.