feat(cli): support array output (#3217)
<!-- Thank you for contributing! -->

### Description

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->

close https://github.com/rolldown/rolldown/issues/3200. 

The previous implement is incorrect for array option. Here is summary
for `watch-cli` and `build-cli` at now.

# `build-cli` 
-  normalized option to flat it
-  run each build at sequential
-  print log
# `watch-cli`
The `WatchOptions` also need to support array output if we support it at
`RolldownOptions`, it is corresponding at rollup.
- normalized option to flat it 
- create bundlers
- start notify watcher
- run each build(watching files) + print log
- waiting changes