feat: support PluginContext emitChunk (#3327)
<!-- Thank you for contributing! -->

### Description

The `PluginContext#emitChunk` is used to add an extra entry point at
rollup. Its special feature is that you can specify the filename of the
entry chunk. It generates a specified entry name at the
module-federation remote. Here is an example. We need to generate a
chunk with `filename: 'remoteEntry.js'`.

``` .js
{
    filename: 'remoteEntry.js',
    exposes: { './App': './src/App', }
}
```

It is a children method of `emitFile` with `type: 'chunk'` at rollup,
ref https://rollupjs.org/plugin-development/#this-emitfile. But here is
the difficult to make the method to sync. So here add an async
`emitChunk` method at the pr, i will try to make it to sync at the
future.

The rollup generate specified chunk name implementation is add the
`fileName` to `Moudle#chunkFileNames` and use it generate chunk at
https://github.com/rollup/rollup/blob/master/src/Chunk.ts#L420
<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->