fix: force cjs wrapper to be included in the output (#3400)
<!-- Thank you for contributing! -->

### Description

Closes https://github.com/rolldown/rolldown/issues/3367.

It's a problem due to https://github.com/rolldown/rolldown/pull/3343.

- Since `import 'cjs'` is consider removed in the final output, so it
doesn't reference the `wrap_ref` anymore. It causes the `StmtInfo` that
owns `wrap_ref` being not included in the treeshaking phase so that the
`wrapping` behavior is gone.
- This PR, by making the `StmtInfo` that owns `wrap_ref` having side
effect, fix the issue.



<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->