feat: support function for assetFilenames (#3397)
<!-- Thank you for contributing! -->

### Description

The `assetFilenames` is used at ecosystem-ci for `nuxt`, the pr is
intended to support it, relate to
https://github.com/nuxt/nuxt/issues/30654.

The `emitFile` is sync method, but call `assetFilenames` function is
async, if using `block_on` for `async emitFile` always hang at call
`assetFilenames` function, because the js side `emitFile` is not
finished. So here calculate the `assetFilenames` at js side to avoid the
issue.

close https://github.com/rolldown/rolldown/issues/3253.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->