fix: escape import/export module name and json key (#3458)
<!-- Thank you for contributing! -->

### Description

- Closes https://github.com/rolldown/rolldown/issues/3456

I used `serde_json::to_string` for escaping json string. I think this
should get the job done, but I'm not sure if the extra dependency should
be avoided in rolldown util.
(I think what this need is just a small escaping utility
https://github.com/serde-rs/json/blob/1174c5f57db44c26460951b525c6ede50984b655/src/ser.rs#L2063,
so copying this code might be an option? :thinking: )