feat(node): allow `sourceRoot: null` in sourcemaps (#3694)
<!-- Thank you for contributing! -->

### Description
According to the spec, `sourceRoot: null` is not valid, but some tools
returns a sourcemap with it. In that case, napi-rs outputs an error
which is difficult to understand by users ("Value is non of these types
`String`, `BindingJsonSourcemap`").
In this PR, I added a code to convert null into undefined to skip that
error.
Note that if `sourceRoot: null` is included in a string sourcemap, it
will be converted to None by serde-json.

Alternatively, we can throw a more friendly error, but I guess we don't
need to be that strict for the input. (Our output should conform the
spec though)

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->