feat(node/wasi): add private `@rolldown/wasi` to prepare to distribute wasm binary in a friendly way (#3925)
<!-- Thank you for contributing! -->

### Description

This PR introduces a standalone package `@rolldown/wasi` to distribute
rolldown wasm binary in a friendly.

However, we sill keep the old way that's using
`@rolldown/binding-wasm32-wasi` package to let the package manager to
install rollldown with wasm on demand and automatically.

`@rolldown/wasi` gives a solution for thoese who want to use rolldown
wasm only.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->