feat: optimize dep unchanged (#3974)
<!-- Thank you for contributing! -->

### Description
This pull request can ensure that if a module is modified without
changing its dependency, only that module will be rebuilt. Before that,
the sub-dependency graph of that module will be appended to the rebuild
queue.
1. I merged `intermediate.importers` into `cache. importers` because the
`fetch_modules` will access the module even if the module is not
changed, before It is ok because the subgraph will be always rebuild,
https://github.com/rolldown/rolldown/blob/4092483d2b287e0f4619544275b2063c8f309fd1/crates/rolldown/src/module_loader/module_loader.rs?plain=1#L409-L413,
the `idx` will be always allocated before access. Now, since the idx
will not always be allocated if the dependency is not changed, if we use
`HybridIndexVec`, it will panic at runtime.
2. Another notable I passed the owned cache instead of
`cache.module_id_to_idx`, this is because we now have two global
maintained field instead of one, maybe more in the future. Passing them
one by one may make the function signature hard to read in the future,
Now the cache has different semantics in different context,
   - In full build, it is almost identical as before.
- In HMR and incremental rebuild, it has some globally maintained data
structure we could modify on the fly during the module loading.
Taking ownership is fine. The program will immediately panic if you
don't set it back when you use it, and it is only used in a few places.
I don't use mutable references, because that will propagate the lifetime
marking everywhere.
<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->