fix!: remove glob filter support for resolveId hook (#3976)
<!-- Thank you for contributing! -->

### Description

Removed the string (glob) id filter support for resolveId hook.
Globs are meant to be matched against absolute path. On the other hand,
raw ids (specifiers) can be relative paths. The semantics of matching
relative paths with globs are not clear and is confusing.

This PR only removes the ability from the types as that was easier and
it would not break existing plugins (if exists).

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->