chore(infra): tweak `pnpm run` ordering (#4043)
<!-- Thank you for contributing! -->

### Description

This PR is improvement for `pnpm run` ordering.
ref:
https://github.com/rolldown/rolldown/pull/4036#discussion_r2030783133

The order was not the same as `npm-run-all2`, so I'm trying to make it
so with this PR.

And, I've just swtiched to `--sequential` option from
`--workspace-concurrency=1`.
We can understand this option name better than the previous
`--workspace-concurrency=1`.
 
<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->