feat(debug): set up event-based tracking system for bundler actions (#4094)
<!-- Thank you for contributing! -->

### Description

tl;dr: This PR just use `tracing` crate to record events and write down
data to the disk in JSON format.

This PR is a rethinking of
https://github.com/rolldown/rolldown/pull/4049.

I dropped the database part and went with a `.json` file for fast
prototyping. This deosn't mean we give up the database approach. It's
just that the database approach need to be built on top of this PR.

#4049 is a quite DX disaster for rolldown developer. Defining a `action`
and insert it to the database request a string of manual work in rust.
And yes they could be solved by defining macros or meta programing
methods with increasing the mataining overhead.

#4049 got dropped due to its bad DX rather than choosing database as the
storeage.

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->