fix: deconfilct chunk name after name generated (#4191)
<!-- Thank you for contributing! -->

### Description
close https://github.com/rolldown/rolldown/issues/4177.

Here follow rollup deconfilct chunk name after name generated, it is
used with `bundle` as `used_name_counts`, so here merge the js/css to
one map.ref
https://github.com/rollup/rollup/blob/master/src/Chunk.ts#L588

Extract it to an isolated function because the asset also could be using
it.
<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->