refactor: processing external module in tokio thread instead of main thread (#4329)
<!-- Thank you for contributing! -->

### Description
#### Context
1. https://github.com/rolldown/rolldown/pull/4305 breaks the wasi build,

![image](https://github.com/user-attachments/assets/0974e239-fe9a-493e-befa-907f0687f53f)
I guess it is due to the fact that I changed `try_spawn_task` to
`async`,whichs is requir.d,
Because `moduleSideEffects` maybe a `jsFunction` call, so it needs to be
converted to async, since async is contagious, the caller also needs to
be converted to `async`.

Now this pr tries to solve this problem without converting
`try_spawn_task` to async, which moves `externalModuole` processing to
another tokio thread.
<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->