chore: remove `preinstall` script (#4331)
<!-- Thank you for contributing! -->

### Description

<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->

Remove the `preinstall` script in the rolldown package, otherwise, the
following warning message will appear when using pnpm to install
rolldown, and relevant configuration is required.


![image](https://github.com/user-attachments/assets/64cec598-f180-47dc-8993-ce0dc1ae5d9a)

In fact, this script has no logical modification, it is only used to
verify the package management tool during development, so I think it can
be removed. The corresponding script can be included in the package.json
file in the project root directory.