refactor: add MarkModuleLoaded option to resolve module load promise (#4356)
<!-- Thank you for contributing! -->

### Description
close https://github.com/rolldown/rolldown/issues/4341

The issue happened if the js side load module, but it already enter the
module task which not finished. Here maybe couldn't be notify in
parallel.

So here add the `MarkModuleLoaded` option to move the related logic to
js side, it could avoid the rust parallel issue.

related:
https://github.com/sxzz/rolldown-plugin-dts/actions/runs/14703765236
<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->