feat: implement filter expr filtering (#4346)
<!-- Thank you for contributing! -->

### Description
The limitation of current hook filter: 
1. You can't specify test `include` before `exclude`, here is an
example:
```js

function hookHandler(id) {
  if (aRegexp.test(id)) {
    executeSomething();
  }
  if (bRegexp.test(id)) {
    return;
  }
}
```
You may write such filter: 
```js
{
  include: [aRegexp],
  exclude: [bRegexp]
}
```
This is correct if and only if `aRegexp` doesn't intersect with
`bRegexp` because `exclude` is always executed before `include` and will
return false if it matches.
real-world example: 

https://github.com/vitejs/rolldown-vite/blob/91a494cb70c0fe21222261878770b23500e9ef41/packages/vite/src/node/plugins/asset.ts?plain=1#L199-L213
2. Express `or` pattern with multiple different field in transform
filter with `include`, e.g.
```js
function hookHandler(id, moduleType) {
  if (moduleType == "js" || /\.js$/.test(id)) {
    doSomething();
  }
}
``` 
You may write filter like: 
```js
{
  moduleType: ['js'],
  id: /\.js$/
}
```
But this did not work at all, copy from documentation
> If multiple filter properties are specified, the filter matches when
all of the specified properties match. In other words, if even one
property fails to match, it is excluded, regardless of the other
properties.

Another real-world example would be: 

https://github.com/vitejs/rolldown-vite/blob/fef84b75dbb35a6ec27debdc0dced1d0f1250eb8/packages/vite/src/node/plugins/importAnalysisBuild.ts?plain=1#L242-L244
3. Don't support preprocessor: for example: 
it is commonly in vite to `cleanUrl` before filter the id: 
-
https://github.com/vitejs/vite-plugin-react/blob/7517103485081b26004e79f169efdd2d12a60946/packages/plugin-react/src/index.ts#L189
-
https://github.com/vitejs/rolldown-vite/blob/91a494cb70c0fe21222261878770b23500e9ef41/packages/vite/src/node/plugins/asset.ts?plain=1#L211
But for now, you need to convert it by your self, this is how vite did:

https://github.com/vitejs/rolldown-vite/pull/118/files#diff-91776e7c6039d23a070162f02a69cd46046a2095bd5ecb384ae9e27f2ea5288fR185-R191
Another example:
https://github.com/sveltejs/vite-plugin-svelte/blob/3589433cd19464c484f560516d41e670e5d40710/packages/vite-plugin-svelte/src/index.js#L166-L168,
It would be harder to write a regular expression to represent that
filter logic.
  
Another benefit of `filterExpr` is the way you write `filterExpr` is
more likely you write javascript, e.g.

```js
  if (aReg.test(id) && bReg.test(id)) {
    return;
  }

  if (cReg.test(code) && !dReg.test(code)) {
    doSomething();
  }

```

the related `filterExpr` would be: 

```js
filter: {
  custom: [
    exclude(and(id(aReg), id(bReg))),
    include(and(code(cReg), not(code(dReg))))
  ]
}
```
They are isomorphic in some sense!
<!-- Please insert your description here and provide especially info
about the "what" this PR is solving -->